Consolidate deprecation tests
